requirements.txt
customtkinter>=5.2.0
cryptography>=41.0.0
argon2-cffi>=23.1.0
pyperclip>=1.8.2
pyotp>=2.9.0
qrcode>=7.4.2
pillow>=10.0.0
mnemonic>=0.20
